pyproject.toml
[project]
name = "tg-bot-vpn"
version = "0.2.0"
description = "Telegram VPN configuration bot with WireGuard support"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "aiogram>=3.0.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/inasekin/tg-bot-vpn"
Repository = "https://github.com/inasekin/tg-bot-vpn.git"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W"]
